refactor: #450 피드백 반영
- 요청 DTO 생성자 접근 제어자를 변경(public -> private)하고,
응답 DTO 는 필요한 필드만 생성해서 내보내기 위해 불변으로 설정했다.
- 테스트 코드를 위한 Line.sections() 메서드 제거하고
다른 방식으로 검증한다
