+++ src/main/java/nextstep/subway/applicaion/dto/SectionRequest.java
@@ -5,7 +5,7 @@ public class SectionRequest {
     private Long downStationId;
     private int distance;
 
-    public SectionRequest() {
+    private SectionRequest() {
     }
 
     public SectionRequest(Long upStationId, Long downStationId, int distance) {
