+++ src/main/java/nextstep/subway/applicaion/dto/StationRequest.java
@@ -3,6 +3,9 @@ package nextstep.subway.applicaion.dto;
 public class StationRequest {
     private String name;
 
+    private StationRequest() {
+    }
+
     public String getName() {
         return name;
     }
