+++ src/main/java/nextstep/subway/applicaion/dto/StationResponse.java
@@ -3,11 +3,8 @@ package nextstep.subway.applicaion.dto;
 import nextstep.subway.domain.Station;
 
 public class StationResponse {
-    private Long id;
-    private String name;
-
-    public StationResponse() {
-    }
+    private final Long id;
+    private final String name;
 
     public StationResponse(Long id, String name) {
         this.id = id;
